Revert VgaSDL -> Vga change from master, because we want to keep the ability to switch backends.
