+++ include/Makefile.am
@@ -187,5 +187,6 @@ noinst_HEADERS = \
 	surface_base.h \
 	surface_sdl.h \
 	version.h \
+	vga_sdl.h \
 	vga_util.h \
 	wav_stream.h
